Resource based authorization

+++ Restaurants.Application/Dishes/Commands/DeleteDishesForRestaurant/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,8 +1,10 @@
 ﻿using AutoMapper;
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishesForRestaurant;
@@ -12,6 +14,7 @@ public class DeleteDishesForRestaurantCommandHandler
 {
     private readonly IRestaurantsRepository _restaurantsRepository;
     private readonly IDishesRepository _dishesRepository;
+    private readonly IRestaurantAuthorizationService _restaurantAuthorizationService;
 
     private readonly ILogger<DeleteDishesForRestaurantCommandHandler> _logger;
     private readonly IMapper _mapper;
@@ -19,12 +22,14 @@ public class DeleteDishesForRestaurantCommandHandler
     public DeleteDishesForRestaurantCommandHandler(
         IRestaurantsRepository restaurantsRepository,
         IDishesRepository dishesRepository,
+        IRestaurantAuthorizationService restaurantAuthorizationService,
         ILogger<DeleteDishesForRestaurantCommandHandler> logger,
         IMapper mapper
     )
     {
         _restaurantsRepository = restaurantsRepository;
         _dishesRepository = dishesRepository;
+        _restaurantAuthorizationService = restaurantAuthorizationService;
 
         _logger = logger;
         _mapper = mapper;
@@ -46,6 +51,9 @@ public class DeleteDishesForRestaurantCommandHandler
                 nameof(Restaurant),
                 request.RestaurantId.ToString()
             );
+
+        if (!_restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Update))
+            throw new ForbidException();
 
         await _dishesRepository.Delete(restaurant.Dishes);
     }
